requirements.txt
fastapi[standard]==0.115.6
pydantic==2.10.3
tortoise-orm==0.22.1
tortoise-orm[sqlite]==0.22.1
aerich==0.8.0
aiosqlite==0.20.0
nicegui==2.8.1